+++ spec/main-spec.js
@@ -1,98 +1,57 @@
-'use babel'
+/* @flow */
 
-import promisify from '..'
+import FS from 'fs'
+import { it } from 'jasmine-fix'
+import { promisify, promisifyAll } from '../src'
 
 describe('Promisify', function() {
-  it('succeeds properly', function() {
-    const callback = function(arg1, arg2) {
+  it('resolves properly', async function() {
+    const promisified = promisify(function(arg1, arg2) {
       expect(arg1).toBe('something')
-      arg2(null, 'wow')
-    }
-    waitsForPromise(function() {
-      return promisify(callback)('something').then(function(retVal) {
-        expect(retVal).toBe(retVal)
-      })
+      arg2(null, 'Doge')
     })
+    expect(await promisified('something')).toBe('Doge')
   })
-
-  it('throws properly', function() {
-    const callback = function(arg1, arg2) {
+  it('throws properly', async function() {
+    const promisified = promisify(function(arg1, arg2) {
       expect(arg1).toBe('something')
-      arg2('wow')
-    }
-    waitsForPromise(function() {
-      return promisify(callback)('something').then(function() {
-        expect(false).toBe(true)
-      }, function(error) {
-        expect(error).toBe('wow')
-      })
+      arg2(new Error('Much Much Doge'))
     })
+    try {
+      await promisified('something')
+      expect(false).toBe(true)
+    } catch (error) {
+      expect(error.message).toBe('Much Much Doge')
+    }
   })
-  it('works when Array.from is available', function() {
-    const _ = Array.from
-    Array.from = function(array) {
-      return Array.prototype.slice.call(array)
-    }
-    const callback = function(arg1, arg2) {
-      expect(arg1).toBe('something')
-      arg2(null, 'wow')
-    }
-    waitsForPromise(function() {
-      return promisify(callback)('something').then(function(retVal) {
-        expect(retVal).toBe(retVal)
-        Array.from = _
-      })
+  describe('throwError as false', function() {
+    it('returns normal output when resolved', async function() {
+      const promisified = promisify(function(arg1, arg2) {
+        expect(arg1).toBe('something')
+        arg2(null, 'Doge')
+      }, false)
+      expect(await promisified('something')).toBe('Doge')
     })
-  })
-  it('works when Array.from is not available', function() {
-    const _ = Array.from
-    Array.from = null
-    const callback = function(arg1, arg2) {
-      expect(arg1).toBe('something')
-      arg2(null, 'wow')
-    }
-    waitsForPromise(function() {
-      return promisify(callback)('something').then(function(retVal) {
-        expect(retVal).toBe(retVal)
-        Array.from = _
-      })
+    it('returns false when rejeted', async function() {
+      const promisified = promisify(function(arg1, arg2) {
+        expect(arg1).toBe('something')
+        arg2(new Error('Much Much Doge'))
+      }, false)
+      expect(await promisified('something')).toBe(false)
     })
-  })
-  it('has an option to disable the throw behavior', function() {
-    function callback(beep, callback) {
-      if (beep === true) {
-        callback(null, true)
-      }
-      throw new Error('Hello')
-    }
-    waitsForPromise(function() {
-      return promisify(callback, false)(true).then(function(result) {
-        expect(result).toBe(true)
-      })
-    })
-    waitsForPromise(function() {
-      return promisify(callback, false)().then(function(result) {
-        expect(result).toBe(null)
-      })
+    it('returns true if resolved with undefined', async function() {
+      const promisified = promisify(function(arg1, arg2) {
+        expect(arg1).toBe('something')
+        arg2(null)
+      }, false)
+      expect(await promisified('something')).toBe(true)
     })
   })
 
   describe('promisifyAll', function() {
-    it('promifies all the functions available on that object', function() {
-      const promisifiedFS = promisify.promisifyAll(require('fs'))
-      waitsForPromise(function() {
-        return promisifiedFS.readFile(__filename).then(function(contents) {
-          return Buffer.isBuffer(contents)
-        })
-      })
-      waitsForPromise(function() {
-        return promisifiedFS.realpath(__filename).then(function(realpath) {
-          expect(realpath).toBe(__filename)
-        })
-      })
-    })
-    it('preserves all the non-function stuff', function() {
-      expect(promisify.promisifyAll({ a: true })).toEqual({ a: true })
+    it('creates async functions and returns similar object', async function() {
+      const promisedFS = promisifyAll(FS)
+      expect(Buffer.isBuffer(await promisedFS.readFileAsync(__filename))).toBe(true)
     })
   })
 })